# Switch to using Integer.digits/2 and add some docs and specs