--- a/lib/luhn.ex
+++ b/lib/luhn.ex
@@ -29,26 +29,28 @@ defmodule Luhn do
 
   def checksum(number, base \\ 10)
 
-  @spec checksum(integer, 2..36) :: integer
-  def checksum(number, base) when is_integer(number) do
+  @spec checksum(binary, 2..36) :: integer
+  def checksum(number, base) when is_binary(number) do
     number
-    |> Integer.to_string(base)
+    |> String.to_integer(base)
     |> checksum(base)
   end
 
-  @spec checksum(String.t, 2..36) :: integer
+  @spec checksum(integer, 2..36) :: integer
   def checksum(number, base) do
     number
-    |> String.split("", trim: true)
-    |> Enum.reduce([], fn(n, acc) -> [String.to_integer(n, base)|acc] end)
+    |> Integer.digits(base)
+    |> Enum.reverse
     |> double(base, 0)
     |> rem(base)
   end
 
+  @spec double([integer, ...], integer, integer) :: integer
   def double([], _, acc), do: acc
   def double([x], _, acc), do: x + acc
   def double([x,y|tail], base, acc), do: double(tail, base, acc + x + sum(y * 2, base))
 
+  @spec sum(integer, integer) :: integer
   defp sum(number, base) when number >= base, do: number - base + 1
   defp sum(number, _), do: number
 end