--- a/lib/luhn.ex
+++ b/lib/luhn.ex
@@ -5,6 +5,8 @@ defmodule Luhn do
   Credit card numbers may be of arbitrary length and in arbitrary base.
   """
 
+  defguardp valid_base(base) when base >= 2
+
   @doc """
   Evaluates a given credit card number for its validity, with an optionally provided base.
 
@@ -21,23 +23,23 @@ defmodule Luhn do
       iex(3)> Luhn.valid?(0x1580BB2EA8875, 16)
       true
   """
-  @spec valid?(number :: integer | String.t, base :: 2..36) :: boolean
-  def valid?(number, base \\ 10) when base >= 2 do
+  @spec valid?(number :: integer | String.t, base :: integer) :: boolean
+  def valid?(number, base \\ 10) when valid_base(base) do
     checksum(number, base)
     |> Kernel.==(0)
   end
 
   def checksum(number, base \\ 10)
 
-  @spec checksum(binary, 2..36) :: integer
-  def checksum(number, base) when is_binary(number) do
+  @spec checksum(binary, integer) :: integer
+  def checksum(number, base) when is_binary(number) and valid_base(base) do
     number
     |> String.to_integer(base)
     |> checksum(base)
   end
 
-  @spec checksum(integer, 2..36) :: integer
-  def checksum(number, base) do
+  @spec checksum(integer, integer) :: integer
+  def checksum(number, base) when valid_base(base) do
     number
     |> Integer.digits(base)
     |> Enum.reverse
@@ -53,4 +55,5 @@ defmodule Luhn do
   @spec sum(integer, integer) :: integer
   defp sum(number, base) when number >= base, do: number - base + 1
   defp sum(number, _), do: number
+
 end