I've been using the Luhn algorithim as an interview question for a few years now, and have found some small efficiency gains that can be made over the standard string splitting version, with regards to elixir.

If you use [`Integer.digits/2`](https://hexdocs.pm/elixir/Integer.html#digits/2) instead of `String.split/3`, you get a non-insubstantial performance boost. This, in addition to the use of Enum.reverse instead of a reducer, makes this variant over 3.5x more performant.

Additionally, I added and updated some typespecs and added some documentation blocks

## Benchmarks
Old variant:
```
Settings:
  duration:      1.0 s

## BasicBench
[12:06:30] 1/1: luhn

Finished in 2.03 seconds

## BasicBench
luhn      500000   3.18 µs/op
```

This PR:
```
Settings:
  duration:      1.0 s

## BasicBench
[12:07:03] 1/1: luhn

Finished in 9.99 seconds

## BasicBench
luhn    10000000   0.90 µs/op
```