@ma2gedev 2019-08-18
Thank you for your pull request containing wide range improvements. 

I've confirmed performance improvement on my laptop. It got the following results.

Old code:

```
Settings:
  duration:      1.0 s

## BasicBench
[11:20:12] 1/1: luhn

Finished in 2.15 seconds

## BasicBench
luhn      500000   3.32 µs/op
```

Your code:

```
Settings:
  duration:      1.0 s

## BasicBench
[11:19:50] 1/1: luhn

Finished in 10.57 seconds

## BasicBench
luhn    10000000   0.95 µs/op
```

I will merge it. Thank you ❤️