@ma2gedev 2019-08-18
@paradox460 I just published 0.3.3 containing this PR. Please check it! Thank you 👍 
https://hex.pm/packages/luhn/0.3.3